[#32] refactor: 분기 처리
